requirements.txt
pre-commit
cityscapesScripts
yapf == 0.26.0
flake8
pyyaml >= 5.1
visualdl >= 2.0.0
opencv-python
tqdm
filelock
scipy
prettytable
paddlepaddle-gpu==2.1.2